Rename BBN SimpleL2Learning so it can be extended
Python does not allow module names to contain dashes
